Renamed library

+++ examples/load_colour_f32.rs
@@ -1,10 +1,10 @@
-use ndarray_images::ImageRGB;
+use photo::ImageRGB;
 
 const INPUT_DIR: &str = "input";
 const IMAGE_NAME: &str = "colour-f32.png";
 
 fn main() {
     let filepath = format!("{}/{}", INPUT_DIR, IMAGE_NAME);
-    let image = ImageRGB::<u8>::load(filepath).expect("Failed to load image");
+    let image = ImageRGB::<f32>::load(filepath).expect("Failed to load image");
     println!("{}", image);
 }
